requirements.txt
streamlit>=1.33,<2
pandas>=2.0,<3
numpy>=1.26,<2.2
scikit-learn>=1.4,<1.7
joblib>=1.3,<2
tensorflow-cpu>=2.16,<2.22
